+++ tests/utils/creditNotePreview.js
@@ -1,0 +1,114 @@
+const starkinfra = require('../../index.js');
+const random = require('./random');
+const { TaxIdGenerator } = require('./taxIdGenerator.js');
+
+const generatorsByType = {
+    'sac': generateSacPreview,
+    'price': generatePricePreview,
+    'american': generateAmericanPreview,
+    'bullet': generateBulletPreview,
+    'custom': generateCustomPreview,
+}
+
+function choice(options=[]){
+    return options[random.randomInt(0, options.length - 1)]
+}
+
+function _choiceBetween(option, chance=0.5){
+    return Math.random() > chance ? option : undefined
+}
+
+exports.generateInvoices = (days = random.randomInt(1, 90), amount=100) => {
+    return new starkinfra.creditNote.Invoice({
+        amount: amount,
+        due: random.futureDate(days)
+    })
+}
+
+exports.generateInvoicePreview = (n=1, amount=random.randomInt(1, 100000)) => {
+    return [...Array(n)].map((item, index) => generateInvoices(days=(index + 1) * 30, Math.round(amount / n)))
+}
+
+exports.generateSacPreview = () => {
+    return new starkinfra.CreditNotePreview({
+        taxId: TaxIdGenerator.taxId(),
+        type: "sac",
+        nominalAmount: random.randomInt(1, 100000),
+        rebateAmount: _choiceBetween(random.randomInt(1, 1000)),
+        nominalInterest: random.randomFloat(0, 4.99),
+        scheduled: random.futureDate(days=random.randomInt(11, 20)),
+        initialDue: random.futureDate(days=random.randomInt(30, 40)),
+        initialAmount: random.randomInt(1, 9999),
+        interval: _choiceBetween(choice(["month", "year"]))
+    })
+}
+
+exports.generatePricePreview = () => {
+    return new starkinfra.CreditNotePreview({
+        taxId: TaxIdGenerator.taxId(),
+        type: "price",
+        nominalAmount: random.randomInt(1, 100000),
+        rebateAmount: _choiceBetween(random.randomInt(1, 1000)),
+        nominalInterest: random.randomFloat(0, 4.99),
+        scheduled: random.futureDate(days=random.randomInt(11, 20)),
+        initialDue: random.futureDate(days=random.randomInt(30, 40)),
+        initialAmount: random.randomInt(1, 9999),
+        interval: _choiceBetween(choice(["month", "year"]))
+    })
+}
+
+exports.generateAmericanPreview = () => {
+    return new starkinfra.CreditNotePreview({
+        taxId: TaxIdGenerator.taxId(),
+        type: "american",
+        nominalAmount: random.randomInt(1, 100000),
+        rebateAmount: _choiceBetween(random.randomInt(1, 1000)),
+        nominalInterest: random.randomFloat(0, 4.99),
+        scheduled: random.futureDate(days=random.randomInt(11, 20)),
+        initialDue: random.futureDate(days=random.randomInt(30, 40)),
+        count: random.randomInt(1, 100),
+        interval: _choiceBetween(choice(["month", "year"]))
+    })
+}
+
+exports.generateBulletPreview = () => {
+    return new starkinfra.CreditNotePreview({
+        taxId: TaxIdGenerator.taxId(),
+        type: "bullet",
+        nominalAmount: random.randomInt(1, 100000),
+        rebateAmount: _choiceBetween(random.randomInt(1, 1000)),
+        nominalInterest: random.randomFloat(0.1, 4.99),
+        scheduled: random.futureDate(days=random.randomInt(11, 20)),
+        initialDue: random.futureDate(days=random.randomInt(30, 40))
+    })
+}
+
+exports.generateCustomPreview = () => {
+    var amount = random.randomInt(1, 100000)
+    return new starkinfra.CreditNotePreview({
+        taxId: TaxIdGenerator.taxId(),
+        type: "custom",
+        nominalAmount: amount,
+        scheduled: random.futureDate(days=random.randomInt(11, 20)),
+        rebateAmount: _choiceBetween(random.randomInt(1, 1000)),
+        invoices: generateInvoicePreview(n=random.randomInt(1, 10), amount)
+    })
+}
+
+exports.generateRandomPreview = () => {
+    return choice([
+        generateSacPreview,
+        generatePricePreview,
+        generateAmericanPreview,
+        generateBulletPreview,
+        generateCustomPreview
+    ]).call()
+}
+
+exports.getPreviewExample = (type) => {
+    return type ? generatorsByType[type].call() : generateRandomPreview()
+}
+
+exports.getPreviewExamples = (n=10, type) => {
+    return type ? [...Array(n)].map(() => getPreviewExample(type)) : [...Array(n)].map(() => getPreviewExample())
+}